format tightening for INSERT
git-svn-id: svn://svn.open-ils.org/ILS/trunk@7526 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ Open-ILS/src/c-apps/oils_cstore.c
@@ -774,10 +774,10 @@ jsonObject* doCreate(osrfMethodContext* ctx, int* err ) {
 			
 		} else if ( !strcmp(osrfHashGet(field, "primitive"), "number") ) {
 			if ( !strcmp(osrfHashGet(field, "datatype"), "INT8") ) {
-				buffer_fadd( val_buf, "%lld", atol(value) );
+				buffer_fadd( val_buf, "%lld", atoll(value) );
 				
 			} else if ( !strcmp(osrfHashGet(field, "datatype"), "INT") ) {
-				buffer_fadd( val_buf, "%ld", atoll(value) );
+				buffer_fadd( val_buf, "%d", atoi(value) );
 				
 			} else if ( !strcmp(osrfHashGet(field, "datatype"), "NUMERIC") ) {
 				buffer_fadd( val_buf, "%f", atof(value) );
